[Feature][Core][Spark3] Preapre spark3 develop
