+++ seatunnel-translation/seatunnel-translation-spark/seatunnel-translation-spark-2.4/pom.xml
@@ -26,27 +26,28 @@
     <artifactId>seatunnel-translation-spark-2.4</artifactId>
 
     <dependencies>
-        <dependency>
-            <groupId>org.apache.seatunnel</groupId>
-            <artifactId>seatunnel-translation-spark-common</artifactId>
-            <version>${project.version}</version>
-        </dependency>
 
-        <!--spark-->
         <dependency>
             <groupId>org.apache.spark</groupId>
             <artifactId>spark-streaming_${scala.binary.version}</artifactId>
+            <version>${spark.2.4.0.version}</version>
+            <scope>${spark.scope}</scope>
         </dependency>
 
         <dependency>
             <groupId>org.apache.spark</groupId>
             <artifactId>spark-core_${scala.binary.version}</artifactId>
+            <version>${spark.2.4.0.version}</version>
+            <scope>${spark.scope}</scope>
         </dependency>
 
         <dependency>
             <groupId>org.apache.spark</groupId>
             <artifactId>spark-sql_${scala.binary.version}</artifactId>
+            <version>${spark.2.4.0.version}</version>
+            <scope>${spark.scope}</scope>
         </dependency>
+
     </dependencies>
 
 </project>
