+++ seatunnel-translation/seatunnel-translation-spark/seatunnel-translation-spark-3.3/pom.xml
@@ -17,27 +17,41 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>seatunnel-translation</artifactId>
+        <artifactId>seatunnel-translation-spark</artifactId>
         <groupId>org.apache.seatunnel</groupId>
         <version>${revision}</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>seatunnel-translation-spark</artifactId>
-    <packaging>pom</packaging>
-    <modules>
-        <module>seatunnel-translation-spark-2.4</module>
-        <module>seatunnel-translation-spark-3.3</module>
-    </modules>
+    <artifactId>seatunnel-translation-spark-3.3</artifactId>
+
+    <properties>
+        <scala.binary.version>2.12</scala.binary.version>
+    </properties>
 
     <dependencies>
 
         <dependency>
-            <groupId>org.apache.seatunnel</groupId>
-            <artifactId>seatunnel-translation-base</artifactId>
-            <version>${project.version}</version>
+            <groupId>org.apache.spark</groupId>
+            <artifactId>spark-streaming_${scala.binary.version}</artifactId>
+            <version>${spark.3.3.0.version}</version>
+            <scope>${spark.scope}</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>org.apache.spark</groupId>
+            <artifactId>spark-core_${scala.binary.version}</artifactId>
+            <version>${spark.3.3.0.version}</version>
+            <scope>${spark.scope}</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>org.apache.spark</groupId>
+            <artifactId>spark-sql_${scala.binary.version}</artifactId>
+            <version>${spark.3.3.0.version}</version>
+            <scope>${spark.scope}</scope>
         </dependency>
 
     </dependencies>
-    
+
 </project>
